Zap manual repo handling in the preseed and kickstart.
This gets rid of all manual repo handling in the preseed (for Debian
and Ubuntu) and the kickstart (for Centos, RHEL, etc).  It replaces it
with the repo-handling helpers that dr-provision has provided based on
the package-repositories parameter.

In particular, this PR removes the meaning the "local-repo" param used
to have for Debian/Ubuntu installs, since what it was trying to do can
easily be accomplished by setting up package-repositories to point to
the mirror of your choice for any OS install repos.