Add a task and stage and parameters that can be used to enforce that a system is sledgehammer.  While hard to do, there are times that the system can get out of sync with itself and the admin wants to enforce that the system is in sledgehammer.  The primary way this happens is that a Next-Boot-Pxe call failed and the system ended up back in the original local booted operating system.
The default discovery stage will allow for enforcing this
or a new stage can be added to an existing workflow,
enforce-sledgehammer.

Whether the new task should enforce or not is controlled
by the parameter 'sledgehamer/enforce'.  This defaults to
false, but when set to true, the task will fail if not
in sledgehammer.  Additionally, the parameter,
'sledgehammer/reboot-if-not-in-sledgehammer`, when set
to true (not the default), will reboot the system instead
of failing.