feat(whoami): Refactor discovery scripts to use whoami if present
This refactors the discovery and join-up scripts to use a variety
of methods to determine what machine UUID matches the machine they
are running on.  It also refactors the code implementing the discovery
methods out to its own script for ease of ongoing maintenance.

The discovery methods are:

* Whoami
* MAC address matching
* Kernel command line.
* Machine-local saved UUID
* Hostname matching.

Note the lack of address based matching -- experience at one of our
largest customers has taught us that IP based matching cannot be relied upon.

This also refrains from setting the machine Address on machine creation,
and includes a task (that is included in the base discover stage) that uses
the same logic our DHCP server uses to fix up the Address field in the face
of oversubscribed DHCP address ranges.