fix(discoverName): Name generation during discovery was broken.
getent hosts during Sledgehammer startup was not filtering out
localhost entries properly, which could (in combination with other
issues) lead to the system getting a hostname of
`localhost\nlocalhost`, which would wig out the rest of the machine
name generation process.  Fixing thios at the point it was broken
would require a respin of Sledgeahmmer, so in the meantime work around
it in start-up.sh in the discover bootenv.