feat(cent8-sledgehammer): CentOS 8 based Sledgehammer
Rewrite the sledgehammer-builder package to base Sledgehammer on
CentOS 8.

Refactor the sledgehammer-builder content pack to take advantage of
the upcoming bootenv-customize changes, and allow several parts of the
Sledgeahmemr build process to be customized via param modification and
task injection.

At the end of the sledgehammer-build process, both the new
Sledgehammer image and a profile that allows you to easily test it
will be uploaded to dr-provision, assuming you are running on a
dr-provision that is capable of handling bootenv-customize.