feat(sledghammer-customize): Optional Mellanox and Solarflare support.
Add tasks that allow building Sledgeahmmer with up-to-date Mellanox
and/or Solidfire drivers.  These drivers amy be required for network
adapters that do not yet have mainstream kernel support.

They are not included in the default Sledgehammer build path because
using drivers compiled from source is not compatible with UEFI Secure
Boot.