fix(burnin): Tweak the disk stress test task.
Ditch incremental status updates when running badblocks.  We run them
all in parallel in the background, and status messages are kinda
pointless in that mode.

Expose a param that lets badblocks bail early after a set number of errors.
The default is to have it try to scan everything anyways.