Script to create a WinPE based on Windows ADK 10
This script will build a basic WinPE based on the Windows 10 ADK.
It supports adding drivers and additional files to the WinPE image
to enable advanced template execution.
The readme file has been update to include instructions to use the
script.
The shutdown batch file is there to support the remote shutdown and
reboot functionality of the drpcli agent, which is not natively
supported within WinPE.