feat(winpe) Add Windows PE boot environment and build script
tools/windows/build-winpe-iso./ps1 will build a basic WinPE based on
the Windows 10 ADK.  It supports adding drivers and additional files
to the WinPE image to enable advanced template execution.  The readme
file has been update to include instructions to use the script.  The
shutdown batch file is there to support the remote shutdown and reboot
functionality of the drpcli agent, which is not natively supported
within WinPE.

The windows-pe boot environment will boot into that boot environment
and start an agent, which can then run Windows-based tasks on the
system.

Shout out to Aptum, who submitted pretty much all of the contents of
this commit.